Final Project -  Multipl Sclerosis leasion MRI image segemntation
the aim of this study is the segmentation of MRI Image for Multiple sclerosis lesions. this project is a part of the main project to be defined for the severity of lesions. Despite many studies that have worked on this topic and published software and algorithms, there are still some limitations in this process that leads to incorrect measurement of lesions. Some of the developed software is based on manual segmentation by user which could lead to poor reproducibility of the final results. Other algorithms use automatic segmentation pipelines that have more reproducibility in their results but are less accurate in detecting some lesions (e.g. cortical lesions).

In this study, 400 participants with relapsing-remitting multiple sclerosis (RRMS) after ethics approval were scanned using multiple 1.5 tesla MRI machines. All participants underwent a medical examination in order to make sure they didn’t have relapses in the last 6 months and also didn’t undergo steroid medication in the last 6 months. Scan protocol contains 3D FLAIR and T1 weighted images with isotropic 1 mm resolution. The same imaging parameters were used in all MRI scanners in order to maintain image quality, SNR, and CNR for all participants.

Segmentation has been done by deep learning using nibabel, keras, TensorFlow, and open_cv. to start the modeling, preprocessing is a crucial step because of reinforces the image and corrects all alignments, and removes noise to have a precise segmentation. one step after alignment is removing the skull in order to do brain extraction then extract binary mask as a final step in preprocessing.

For the modelling part, Unet with resnet50 as backbone has been used. the reference for this project is calibrated according to article which is mentioned in my project at the end of the codes. the result has been evaluated by dice coefficient(mean IOU for comaparing). inspite of the acuaracy as 1, the dice coefficient is low. because of mask background which is more than 95% black, so in comparing pixel by pixel, accuracy is really near to 1, thugh for region covering, due to have small spot which is the expected segmnetation as a correct result, is low. in order to improve this result, this project is going to progress. I will try to update this project with the better result in near future. Next update would be one of the pivotal step as a preprocessing is image augmentation.

I appreciate to have any idea ar any correction in model and codes letting me know about.

[ ]
Colab paid products - Cancel contracts here