Regression Discontinuity Design (RDD)
Regression Discontinuity Design (RDD) is a key method used to determine a certain effect of an intervention based on comparing de before and after a specific threshold. For this project we used the open source database from the National Institute of health of Quebec. This specific csv files has the number of cases per day as well as deaths, accumulated cases, accumulated deaths, etc; For now we are only going to focus on the dates and the number of cases.
Regression Discontinuity Design (RDD) is a key method used to determine a certain effect of an intervention based on comparing de before and after a specific threshold. For this project we used the open source database from the National Institute of health of Quebec. This specific csv files has the number of cases per day as well as deaths, accumulated cases, accumulated deaths, etc; For now we are only going to focus on the dates and the number of cases.